+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [v7.0-beta.144] - 2021-09-08
+
 * Issue **#2398** : Fix to clear interrupt state for threads used by terminated tasks.
 
 * Issue **#2396** : Add `stroom:pointIsInsideXYPolygon` XSLT function.
@@ -3674,7 +3677,8 @@ Issue **gchq/stroom-expression#22** : Add `typeOf(...)` function to dashboard.
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.143...HEAD
+[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.144...HEAD
+[v7.0-beta.144]: https://github.com/gchq/stroom/compare/v7.0-beta.143...v7.0-beta.144
 [v7.0-beta.143]: https://github.com/gchq/stroom/compare/v7.0-beta.142...v7.0-beta.143
 [v7.0-beta.142]: https://github.com/gchq/stroom/compare/v7.0-beta.141...v7.0-beta.142
 [v7.0-beta.141]: https://github.com/gchq/stroom/compare/v7.0-beta.140...v7.0-beta.141